Split Inet::EndPoint into per-implemention files
#### Problem

This is a step toward #7715 _Virtualize System and Inet interfaces_.

#### Change overview

Move each `Inet::EndPointBasis` implementation into its own file.
Each inherits from `Inet::EndPointBase`, but transitionally each
implementation is statically named `EndPointBasis`, and initialization
continues to be done in `InitEndPointBasis()` rather than a constructor.

#### Testing

CI; no changes to functionality.
