+++ src/inet/inet.gni
@@ -25,4 +25,11 @@ declare_args() {
 
   # Enable TCP endpoint.
   chip_inet_config_enable_tcp_endpoint = true
+
+  # Inet implementation type.
+  if (chip_system_config_use_lwip) {
+    chip_system_config_inet = "LwIP"
+  } else {
+    chip_system_config_inet = "Sockets"
+  }
 }
